802.1x: avoid race condition in secret request
'nmcli -a conn up ...' first registers a secret-agent to NM and then
activates the connection.

If we allow the new connection to autoactivate, it is possible that
during autoactivation NM will ask secrets to the nmcli agent just
registered. After that, the manual activation will fail due to lack of
secrets.

This race can be avoided by disabling autoconnection.

Change-Id: I081f89c9f0edc15dc2c74534c37b6b888394282a