ipv4: fix hostname check @ipv4_ignore_resolveconf_with_ignore_auto_dns_var3
After a static hostname is set, transient one is no more meaningful. In
particular, if a transient hostname different from the static one is
set, "hostnamectl --transient" will show for few seconds the one just
set, then will update it to the static one (while the hostname returned
will always be the static one).
Don't do any assumption on the lasting of this temporary transient
hostname in hostnamectl --transient: remove the check (as it happens
that the returned value during the 5 seconds is already the static one).

Change-Id: I8b393278ed0da5d6e9fc74893fcacbfe5f6fdbc2