general: check writable memory only for RSS changes in stable_mem_consumption test
We cannot predict RSS memory usage by NetworkManager, imagine what
happens if somenthing async, e.g., a connectivity check, triggers the
load of some shared library.
In order to have it somehow more under control, consider RSS memory just
for writable memory: so, we monitor memory we allocate for data
structure that is the one that we really want to monitor.
Please, note that after this change would be reasonable to lower the
threshold for memory size changes.

https://bugzilla.redhat.com/show_bug.cgi?id=1452575

Change-Id: Id9b10bf9795618e46f2320a76d7ae001fe74bf24