vethsetup: enforce write of uuid in original ifcfg host file
If the host image used for the test has no UUID key in the ifcfg file
of the active connection, NetworkManager will generate a new UUID for
that connection when the file is loaded.
As we then save the ifcfg file and restore it changing its name, we
need the UUID inside the ifcfg file otherwise NetworkManager will
generate a new UUID when the restored ifcfg file is read again ("nmcli
connection reload". If we lose the UUID, subsequents "nmcli modify"
commands addressing the old UUID will fail.
Overwrite the original connection name as soon as the UUID is read in
order to write down all the keys related to the connection (including
the UUID) before saving the file.