bond,team: first disconnect slaves and then master
Some tests do this:

      * Disconnect device "nm-bond"
      * Bring "down" connection "bond0.0"
      * Bring "down" connection "bond0.1"

When the master gets disconnected it also disconnects slaves, and so
the other two steps are no-ops.

In some tests we then check that slaves don't auto-activate, but this
can only work if the slaves are explicitly disconnected (which doesn't
happen right now because step 2 and 3 have no effect).