ipv4: change renewal_gw_after_dhcp_outage_for_assumed_var1 on 1.10.1
The purpose of the test is to check that NM keeps renewing the lease
for an external DHCP connection. NM used to do that in the past
because it could happen that after a NM restart the connection
matching didn't work and thus the on-disk connection was not
assumed. In such situation, it was considered better to continue doing
DHCP rather than losing connectivity.

Now NM doesn't do connection matching on daemon restart but uses a
state file to know exactly which connection was active. Therefore
there is no risk that a DHCP connection becomes external. If a
connection is external, we should not touch it anymore.

Update the test to reflect this behavior.

https://bugzilla.redhat.com/show_bug.cgi?id=1518091