general: update @preserve_failed_assumed_connections test
Instead of checking that the device state is "connecting", just check
that the IP addresses is preserved. Until now, when NM found an
external connection with dynamic addresses, it started DHCP (ignoring
the result) and thus the device stayed in the "connecting" state.

Now NM no longer starts DHCP and the device goes to "connected" state
directly.

https://bugzilla.redhat.com/show_bug.cgi?id=1530288