ipv4: change timeouts for @dhcp4_outages_in_various_situation (#154)
NM changed the way in which lease expirations are handled: before it
stopped the client and respawned it 3 times at fixed intervals (2
minutes), while now it just keep the client running. dhclient tries to
renew the lease immediately and then sleep for a random interval
between 300 and 450 seconds.

Therefore, extend the timeout to be sure the client is able to renew
the lease.