ipv6: fix @ipv6_add_static_address_manually_not_active test to wait
The test is supposed to check that NetworkManager does not add a link
local address. It needs to wait longer to ensure that this doesn't
happen. The previous form did this wrong, which can also be seen because
the test still passes against current master of NetworkManager --
although it would be expected to fail. Fix it.

Fixes: e12de017e6668a48490490abdae0f660024bc51d