ipv6: fix @ipv6_add_static_address_manually_not_active to not expect … (#152)
* ipv6: fix @ipv6_add_static_address_manually_not_active to not expect IPv6 link local address

Before https://github.com/NetworkManager/NetworkManager/pull/75 (1.11.1
and older), NetworkManager would also create link local addresses for
externally assumed devices. The previous version of the test correctly
tested that behavior. But I think this is not desired behavior.

Note, that commonly NetworkManager keeps disconnected devices IFF_UP,
with "addrgenmode none". That is a terrible hack, to get carrier change
events for the device without IPv6 (link local) addresses.

When the user starts externally configuring addresses on the interface,
NetworkManager noticies this and generates an externally assumed
connection. In this mode, NetworkManager must no longer interfere with
whatever the user is doing.

This also means, to not re-add the IPv6 link local address, although
that might be quite unexpected to the user. It is also a change in
behavior with 1.11.2.

* updating device in from eth1 to eth3

from now on we use just eth2, eth3, eth10 devices in ipv6 feature. Moving from eth1 to eth3 in ipv6_add_static_address_manually_not_active