dhcp: update tests for client identifier (#153)
* dhcp: update tests for client identifier

We uniformed DHCP client identifier between the dhclient and internal
backends. As a result, the client identifier sent by dhclient
changed. Since tshark doesn't report the client-id when it contains a
non-printable character, use tcpdump instead. Also, add a test for the
internal client.

[1] https://bugzilla.gnome.org/show_bug.cgi?id=793957

* update to new device and naming scheme

eth2 is used instead of eth9 and ethie name was obsoleted in favour to con_ipv4 in ipv4.feature