bridge: test connection.autoconnect-slaves behavior
autoconnect.slaves should forcefully activate all slaves, even if the
device is currently busy with another profile. That was the case already,
but slightly broken with [1].

This adds two tests, one a general test, and the other where the
profile that should be stolen for activation is the slave profile
itself, but activated previously as a non-slave before being modified.
This is the scenario of bug [2].

[1] https://cgit.freedesktop.org/NetworkManager/NetworkManager/commit/?id=0922a177385be188b9c9c8ad39c1068533f5a4b3
[2] https://bugzilla.redhat.com/show_bug.cgi?id=1548265