connection: add more assertions to tmp/repro_1401515.py
For one, iterate the mainloop after making the asynchronous Update2()
request. This way, we don't need to wait unnecessarily long.

And plaster the test script with checks that everything works as
expected. We currently don't have many tests that test libnm via
pygobject. Use the opportunity and make more use of libnm via Python
and see if it all works as epected.