ipv4: fix @dhcp-timeout_default_in_cfg test (#181)
Tests can run with veth devices instead of ethernet; therefore match
also veth device type in the configuration snippet.

While at it change the connection with autoconnect=no, and manually
activate it. Otherwise, we can't be sure that the test succeeds
because the inifinity timeout was honored; for example, it can happen
that DHCP fails after 45 seconds and autoconnection kicks in.