Adjust unreachable gateway route tests (#183)
* ipv4: modify IPv4 routes_not_reachable tests for new behavior

Since version 1.11.3 NM automatically adds a device route to the route
gateway when it is not directly reachable. See [1].

[1] https://cgit.freedesktop.org/NetworkManager/NetworkManager/commit/?id=c9f89cafdf588f443821ccff220283859ab26d1f

* ipv4: adjust @ipv4_route_onsite test to new behavior

Since version 1.11.3 NM automatically adds a device route to the route
gateway when it is not directly reachable. See [1].

[1] https://cgit.freedesktop.org/NetworkManager/NetworkManager/commit/?id=c9f89cafdf588f443821ccff220283859ab26d1f

Therefore, just check that the onlink route is there, not that the
activation fails without "onlink".