nmcli/vlan: adjust test for abbreviating connection setting as "eth"
nmcli supports abbreviations. For example

  nmcli connection "$PROFILE" ethe.mtu 1500
  nmcli connection "$PROFILE" ethernet.mt 500

This is allowed as long as the abbreviation is unambigious.
As also explained in `man nmcli`:

  NOTES
    nmcli accepts abbreviations, as long as they are a unique prefix in the
    set of possible options. As new options get added, these abbreviations
    are not guaranteed to stay unique. For scripting and long term
    compatibility it is therefore strongly advised to spell out the full
    option names.

Anyway, with rh#1335409, we added a new setting "ethtool", hence
the abbreviation "eth" is no longer unambigious:

  $ nmcli connection "$PROFILE" eth.mtu 1500
  Error: invalid or not allowed setting 'eth': 'eth' is ambiguous: ethernet, ethtool.

While such change in behavior is annoying and possibly problematic,
we do them. I personally think, abbreviations are evil and we should
have never used them. But as we have them, it seems better to fail
in face of ambiguity, instead of continuing to resolve "eth" as "ethernet".

Sidenote: `nmcli device s` has no problem with the ambiguity and
resolves to "status", instead of "set" or "show".