version: fix execution for not_with_rhel7_pkg tag
We need to be able to execute tests with the same and not_with_rhel7_pkg
tag. If this tag is granted that tag will be skipped if package is
stock rhel7. The other test is executed instead.
This wasn't true before as the test was selected as the highest
available version but skipped from inside of environment.