devel/fg/libreswan import export rh1626485 (#242)
* vethsetup: fix typos in comments

* libreswan: rework the export test

instead of checking that the exported config matches the sample export
file (which is a bit old now, but should be preserved for checking older
NM-libreswan plugin version), check that the export contains all the
options of the reference file. So, different order is allowed and also
new options are.

Deal also with the change of the "esp" libreswan option: it is now
deprecated in favor of the new "phase2alg" one that we now use when
exporting.

Lastly, do a self check by importing once again the connection but this
time from the newly exported file and assert that the connections
imported from the reference export and the actual export file are
identical.