add strongswan tests
take some libreswan tests and modify them to work under strongswan 
setup. Some libreswan test are not working under strongswan (imports, 
exports, ask for secrets...).

strongswan is not setting the gateway for some reason, maybe something 
missing in configuration (configurable options are at 
https://wiki.strongswan.org/projects/strongswan/wiki/Swanctlconf), will 
investigate later.

In prepare script, interface strongswan2 with different subnet is 
created to test that tunnel works, but did not manage to create test 
for that.