ipv6: fix ipv6_routes_with_src
there was a bug, that src was not visible with `ip -6 r` command, and
this test did not cover it

Also modified the test to cover the bug, where some bytes of destination
route are zeroed.

This two bugs were related to wrong memory structure used for storing
the IPv6 route settings.

https://bugzilla.redhat.com/show_bug.cgi?id=1727193