dns: adjust test @dns_dnsmasq_kill_ratelimit for change in 1.22 (1.21.2-dev)
When dnsmasq process keeps crashing, NetworkManager will restart it, but
also ratelimit the starts.

Before 1.21.2, when ratelimiting hits, NetworkManager would write the real
name servers to /etc/resolv.conf, instead of 127.0.0.1. The aim ws to
workaround the issue. However, in this case something is really wrong
already. When dnsmasq keeps crashing, there is no solution except letting
the user investigate the issue. Just pretending that for a moment all is
good, until we retry starting dnsmasq again, is not helpful.

This will change in 1.21.2 [1]. We of course still ratelimit the
restarts, but we don't bother "fixing" the situation. Yes, name
resolution won't work, but that seems preferably, as it lets the
user become aware of the problem.

[1] https://gitlab.freedesktop.org/NetworkManager/NetworkManager/merge_requests/254