runtest: unify {nmcli,nmtui}/runtest.sh implementations
{nmcli,nmtui}/runtest.sh is mostly duplicated. Unify them and make
them one script.

It's bad to have duplicated code. For one, it tends to diverge. For
example, nmcli/runtest.sh handles "CRASHED_STEP_NAME", nmtui does not.
That does not seem right, crashes should also be handled by the nmtui
runner.

Also, the run scripts delegate to common setup scripts like "prepare/envsetup.sh".
Indeed there is a need for sharing scripts. But because "runtest.sh" was
not shared, we added other scripts, that indeed got shared.
Scripts only get more complicated, the more there are of them.

In a first step, merely make both scripts identical. The places that differ
now check for the test type, and conditionally do one thing or the
other. That means, the places that differ moved closer "down", to where
the difference is.

Only the next patch will merge the scripts. This is to make the diff
clearer and reviewable.