patch python-dbusmock for latest fix
With libnm rework in upcoming 1.22, NMDevice will only consider
the "StateReason" D-Bus property (to both get the "State" and the
reason). python-dbusmock was/is not suitable for that yet.

Patch our local version of python-dbusmock so that this works.

This affects tests with tag "@mock", which currently is only
"@nmcli_device_wifi_with_two_devices"