cli: extend wait timeout for @vlan_create_many_vlans test
Sometimes, the script to add 255 VLANs takes longer than 30 seconds. That means,
the test fails with timeout, although NetworkManager didn't do anything bad
to the externally created interfaces.

One question is, why it takes so long to add a (relative) small number of interfaces.
Note that NetworkManager in such scenarios is slow and shows performance issues. That
we enable level=TRACE logging only makes it works and puts pressure to systemd-journald
as well. But ultimately, I don't know why it sometimes takes that long.