ipv4: add 'ipv4_external_addresses_no_double_routes' test
Add test, that NM does not add routes for addresses added externally
(e.g. via ip command). The kernel/tool should do that, and if so, the
routes are doubled, and so only one of them is removed. For /32 routes,
kernel does not add them, so NM should not either.

https://bugzilla.redhat.com/show_bug.cgi?id=1700415