version_control: add possibility to override installed NM version
this feature is handy when developing bug reproducers and reproducer has 
to fail on affected NM version. That was not possible without change to 
the tags of the test. Now, this change can be made on tested machine, so 
no need to change the tags of the tests.

This also adds possibility of trying how diffrent versions of the same 
tests behave on different NM versions.

There are two ways how to override NM version. 
 * environment variable `NM_VERSION`
 * file `/tmp/nm_version_override`

The first option is of higher priority, it is enough to specify version 
in format `1.2.3`