vethsetup: catch Fedora openstack setup
Fedora in openstack has in-memory connection for single ethernet device. 
Vethsetup disconnects all devices, removes all in-memory connections and 
then searches for device with connection. Since original connection is 
in-memory only, it gets destroyed and so script fails.

The fix is to activate interface, if there is single ethernet device 
down.