dracut: rewrite ip checks to be more stable
remove simple ip_ifname check, as there are ip(4|6)_forever and 
wait_for_ip(4|6)_renew.

`nmcli_con_active` is  dynamically waiting until connection gets 
activated (`--active` switch is not enough, as connection is active even 
while it is getting the lease). Also, make this before the IP checks, to 
be sure that IP is assigned.