keyfile_plugin: vethsetup rework
We need to use /etc/NetworkManager/system-connections/ instead of
/etc/sysconfig/network-scripts/ to store connections. We need to
check which plugin is default and deploy /tmp/nm_plugin_keyfiles if
keyfiles.
We use the same concept for storing backup of eth0 connection in /tmp.
Restore in environment is handled the same way too but with different
path if /tmp/nm_plugin_keyfiles exist.