steps: reformat file with python black
The files "nm{cli,tui}/features/steps/steps.py" are basically
the same and should be unified/merged.

In a first step, reformat them both with python black code formatter.
That results in a unified form that makes it simpler to find the
similarities. Also, the black format is intentionally designed to
minimize the diff of changes, it thus will be a suitable format to
merge the two files.