steps: replace "nmtui/features/steps" by a symlink to nmcli
Avoid the duplication of our python code. We don't need separate
"steps" implementations. Especially because they are mostly similar
(or the same), and because we want that the steps are similar (the same)
for nmcli and nmtui tests.

To avoid that, symlink now the entire "steps" directory from
nmtui/features/steps to nmcli/features/steps.