steps: rename "nmcli/features/steps/{steps.py => nmci_step.py}"
There should be one utility "library" for our tests.
"steps.py" already is that, but the name doesn't seem
great because it's the default file for behave for steps.
But our "steps.py" contains no steps at all. Rename instead
to "nmci_step.py".

Note that next we will also add a new "nmci" python module with
additional helper utilities. So, we will have "nmci" module that
should be usable everywhere, and we will have "nmci_step.py" with
helper code that is only used for implementing the behave steps.

Optimally, helper functions are general purpose and should be added to the
"nmci" module, and not be specific to the behave steps ("nmci_step.py").
That is because we will add pytest tests for the "nmci" module, but not
for "nmci_step.py". Hence, stuff should preferably go to the future "nmci"
module.