steps: always import full "nmci_step" module and fully qualify the utils functions
Let's be clear about where the helper functions come from. That means,
import the entire "nmci_step" module and fully qualify the helpers we
use. It makes it easier to see where this comes from, especially because
the "pexpect" and "subprocess" modules also have command to run a process,
and they tend to have all similar names.