version_control: support version tags for stable versions
Before, we could only specify one "ver+" or "ver-" tag (or both kinds,
to define a range). As a special convenience, we could specify also
one stable branch, and automatically get the next major release covered.
That means, "ver+=1.26.4" means it covered 1.26.4+ but also 1.27+.

Imagine a bugfix or feature gets backported to older versions.
For example, it is now available in @ver+=1.26.6 @ver+=1.28.4 @ver+=1.29.4
That means, it works not on 1.28.2, but it would work on 1.28.6.
Expressing versions for older stable branches (except the ""ver+=1.26.4"
example above) did not work.

Extend the version comparison to account for multiple version tags, in a
way that makes sense and is suitable to also handle old-stable branches.

Note that there is also a change in behavior: ver+1.28 would previously
also accept "1.28.2". Now, this is the same as ver+=1.29. However, no
"ver+1.*" tags are currently used, so this change has no effect in
practice.