rename tests to not contain the word "fail"
When I look at the test results (in the web page), the first thing thing
I do is CTRL-F "fail". That finds a lot of irrelevant occurrences, because
the term "fail" is part of the names of tests.

That is annoying. Rename those tests.

Yes, I know that all the names here are lower case "fail", while I'd need to
search for "Fail" case-sensitive. Still, let's avoid that and rename the
tests.

Some tests were renamed entirely to avoid the term "fail". Other tests,
got a name with the intentional misspelling "faiil".