add dummy test "pass" which (almost) does nothing
When we schedule a job in jenkins to reserve a machine, we
don't want to run all tests. So we have to specify at least
one test. Since in that case we don't care to run any actual test,
we can either pass a garbage name (which seems bad) or lookup
an existing test.

The problem with the existing tests as that they don't have a memorable
name. So, to find them, you really have to grep the sources.

Add a special test "pass", with a name that is easy memorable and can be
used for this purpose.