ipv4: tighten up check for nameserver in @ipv4_dns_resolvconf_symlinked,@ipv4_dns_resolvconf_file
This effectively reverts commit 1a5da593bd7a ('ipv4: dns: wait for nameserver
to be wriitten').

Back then, the workaround was necessary due to rh#1778798. I can no
longer reproduce that old issue, so presumably it was fixed.

Drop the workaround (by setting the wait-time to zero).

Note that this also re-enables the test for older NetworkManager, which
potentially still have the problem. I would like to see that happen, to
understand how this happens. So this change intentionally affects older
versions too, to possibly reproduce the original problem.

https://bugzilla.redhat.com/show_bug.cgi?id=1778798