dracut: add 'dracut_NM_NFS_remote_rootfs_connection[_var2]' tests
Add tests, that NM is able to use persitent profiles saved on remote 
root. This seems to be problematic, if tested on interface used for NFS 
root (even if configured the same in initrd and remote), so another 
interface added (iscsi0), and it changes from DHCP (initrd kernel 
argument) to manual IPv4 (profile saved in remote root).

https://bugzilla.redhat.com/show_bug.cgi?id=1934122#c24