ovs: remove obsolete step from openvswitch_ignore_ovs_network_setup
This step was actually used until 1cfc489c for copying of uuid of
preexisting connection for newly created one. We don't use it since then
nor there is any check of uuid in latter steps, there are no errors when
the test is run without this step so it's removal should be safe.