publish behave logs/ipv6: change port used for PMTU testing to 9000
Port 8080 is used as well-known secondary http port so let's use it
for behave logs publishing. Just one test uses it as well so let's
move this test to a custom port. No other test uses 8080 port (no
other '8080' string is in features/).

Add check to @ipv6_keep_external_addresses test that no dummy interface
is present on the system