ipv4: fix a race in ipv4_method_shared_with_already_running_dnsmasq
dnsmasq instance we start with --bind-interfaces would bind to all
interfaces with an address on startup. The test expects it to bind
before NetworkManager-managed instance of dnsmasq attempts to bind
and end up failing. Given it's started in background when the
interface is not configured yet, it may or may not end up trying
to bind at the right time.

It inded fails occassionally, roughly once in a few hundred runs.

This patch makes sure the interface is configured at the time dnsmasq is
started and gives it hopefully enough time to bind.