ipsec: avoid disabling ipv6 for *swan tests
Long ago, Libreswan's pluto got confused upon seeing tentative IPv6 addresses
and we ended up disabling IPv6 globally for tests that involved it.

We, however, have a test where Libreswan is used along with an IPv6 OpenVPN
network. The afforementioned hack causes the OpenVPN tun device to be
created with IPv6 off, causing the assertion that an IPv6 address is
present to fail.

This used to be okay because NetworkManager ended up setting disable_ipv6=0
for said interface. This changed in NetworkManager/next branch and I
believe we new behavior is perhaps more correct.

Do away with the Libreswan hack, ensuring that the libreswan version we got
is good enough.