openvpn: tolerate link-local IPv6 addresses in IPv4 tests
The VPN interface gets an IPv6 link-local address even when the server
doesn't supply any other IPv6 addresses. Unless IPv6 is disabled, this
is the correct behavior. It was always the case with both NetworkManager
and the standalone openvpn client.

Before NetworkManager/next, NetworkManager wouldn't expose these
addresses on the VPN active connection. Given the address is actually
there, the new behavior is better. It's also very unlikely to cause any
unwanted effects.

Adjust the tests to accomodate for that.