ipv6: add @ipv6_restart_prefixroute
This is a test for an issue, where an externally added address, that
would ordinarily cause a route to its prefix to be automatically added,
would be re-added into kernel with "noprefixroute" attribute upon
daemon restart. That would cause the automatically added route to be
removed.

The problem was fixed with NetworkManager/next merge (into 1.36).

https://bugzilla.redhat.com/show_bug.cgi?id=2010640