Avoid doing a 'pkill $(ps | grep)'
Grepping process listing is fragile. The result typically includes the
grep, and the kill command themselves. While we get rid of the grep
process with a 'grep -v grep', we still include kill, which can,
unpredictably, shoot itself in the foot.

pkill does the filtering for us -- let's use it.