dracut_NM_iSCSI_ibft_table: assert the address is permanent with iBFT
The iBFT contains an address that has been gotten from DHCP, but no
further details. We don't know how was it obtained, and for how long is
it supposed to be valid etc. Because of this, it's not possible to for
nm-initrd-generator to pass it along to NetworkManager to treat it as
DHCP lease.

To be on a safe side, nm-initrd-generator treats it as a permanent address
instead. Given this address is likely required for accessing a network
volume that's critical for operation of a network-booted system, this
is probably a wise idea.

However, when DHCP is run on the iBFT connection, it's likely to assign
the very same address. Prior to NetworkManager 1.36, when the same
address is assigned manually as well as dynamically, its lifetime would
sometimes be that of the DHCP lease and sometimes permanent depending on
various factors.

With NetworkManager/next branch merge, the precedence is clear -- the
manually assigned address allways *wins*. Let's adjust the iBFT test to
reflect that.

https://bugzilla.redhat.com/show_bug.cgi?id=2013921