lib: make sure we check for crash when NM is responsive
Looking at the PID is not sufficient to determine whether NM crashes.
The process could still be around, albeit unresponsive while the dumper
process sucks whatever's left of its miserable core.

If NM responds to a command, we can be sure it either didn't crash yet,
or is finished with is crashy business and a replacement daemon running
instead. A PID check will reveal which one is the case.