prepare: hostapd_wireless: fix logging and wep_key_len
wep_key_len_broadcast and unicast interprets \"5\" as 0, so it is safe 
to omit (it is also computed as strlen). Weird that setting it correctly 
to 5 or 13 does not work, but it is WEP, no one is gonna fix that. 
However, wep_key_len_ must be set in dynwep, otherwise wep tls fails.

Logging set by hostapd argument -ddd causes wpa_debug_level being too 
verbose (0 - verbose debug), we need it 1 - debug, -dd is the only 
switch how to achieve that (no config for that), so we need to have 
loglevel 3 - warning set in config, which gets subtracted to 1 - debug, 
because of -dd argument.

Also, fix typo in wpa_supplicant log (hostaname), do not log with 
timestamps in hostapd (seems messy), and append logs when env NMCI_DEBUG 
is set.