nmci: don't rename the "tags" module during import
Don't rename. It has a name so use it.

The rename was of course done, because "tags" would also be an
appropriate variable name for other uses. So there is a clash.
But if that is the problem, then the module should not have a name
that you want to use for a variable.

Also, this problem is only in nmci itself. Otherwise, we usually should
do the import with the full name "nmci.tags".