ipv6: add test for prefix-delegation with link-local addressing
Add a test for IPv6 prefix delegation when the upstream interface is
not using SLAAC or DHCPv6 addresses. See:

https://github.com/coreos/fedora-coreos-tracker/issues/888

The difference compared to existing test is that NM is not already
using DHCPv6 on testX6, and it must start the DHCPv6 client when the
downstream interface requests a prefix.

Also, since the server doesn't give out the address via DHCPv6, we
compute the route gateway in a different way. In a real scenario the
server would know the address by other means (for example, because the
link-local interface-id was pushed via PPP IPV6CP).