connection: avoid race in checking Metered status
The test @connection_metered_guess_yes had both IPv4 and IPv6 enabled.
The metered status is autodetected based on an DHCPv4 option, so if
IPv6 completed first the device was considered activate, and the test
failed with wrong metered status. Fix that race by checking for a few
seconds.

Also, clone the test. In the new variant @connection_metered_guess_yes_ipv6_disabled,
IPv6 is disabled and we expect the metered status to be ready right away.