nmci: hook nmci.process from context
I think that our "run" commands should be more strict and have a
"tighter" API.

For example:

- don't just pass "*a, **kw" arguments through to subprocess.run() --
  which itself has a large number of non-trivial arguments that affect
  each other.

- don't accept an encoding parameter, when we almost always want to deal
  with UTF-8.

- fail (by default) if the output is not valid UTF-8.

- fail (by default) if anything gets printed to stderr.

We should thus replace nmci.run.run() (and nmci.run.*), and
context.run() (and context.command*()) with functions that are strict
by default.

In step 1, hook yet another set of run() commands to the context. They
can be called as context.process.run() and context.process.run_check()
and are supposed to work the same as process.run() and process.run_check().
The idea is in the future to replace the other variants.