nmci.tags: use `ip link` to add device with non-UTF-8 name
pyroute2 causes a deprecation warning in the unit test due to
distutils (at least, on my machine).

It seems unnecessary anyway. I mean, sure, if we prefer to use
pyroute2 library over ip command line, then lets do that. But currently
the vast majority to code runs ip and not using pyroute2. In this
case, there is no need to use pyroute2, if we don't do it in general.