environment: do not start NM if found stopped
This breaks run_once tests, as the test leaves NM stoped with 
"configure-and-quit" option enabled.

It is useless to start NM in this case. If it was stopped on purpose 
during the test, there must be @restart_if_needed. In case of crash such 
that `systemctl start NetworkManager` is not able to start NM, it is 
usesess either to start it again, before tags cleanup or after crash 
reset (where NM config is cleaned and NM restarted).