nmci.test: move test_black_code_fromatting() to the end
While developing a test, we may often quickly change something and
forget to fix the formatting. We usually care about the formatting
as the last thing, when all other tests pass. Move it to the bottom
of the file, so that it runs last.

It's also the slowest test so far.