gsm_sim: check for IP address and DNS
Check that the right address is configured on the PPP interface and
that the name server is in resolv.conf.

To obtain the DNS via PPP, tweak the command line arguments of the
modem emulator script.

Note that the last check on DNS currently fails on NM main due to a
regression introduce in 1.36. It will be fixed by:

https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/merge_requests/1122