ipv6: adjust @ipv6_check_addr_order test for bugfix
@ipv6_check_addr_order was only recently added, with the purpose
of testing the *current* behavior -- not necessarily the desired
behavior. I wanted first get the test to pass and test what is
happening, not what should be happening.

The IPv6 address order has many issues. One glaring bug that was
introduced in 1.35+ will be fixed by [1].

This reorders the addresses, specifically so that autoconf6 addresses
are before DHCPv6 addresses, which are before manual addresses.

This is then the reverse of what we really would want to have (we would
want to have manual addresses over DHCPv6 over autoconf6 addresses). But
the intermediate fix done by [1] addresses bugs where the code did not do
the correct thing, not because of disagreeing about what it should do,
but because it just was wrong. For the moment, we keep the odd sorting,
but fix the bug [1]. Later, we might want to change behavior and fix the
desired order.

[1] https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/merge_requests/1168

https://gitlab.freedesktop.org/NetworkManager/NetworkManager-ci/-/merge_requests/1015